+++ src/errors/mod.rs
@@ -4,6 +4,7 @@ mod collection_error;
 mod exceeds_capacity;
 mod item_count_mismatch;
 mod one_of2;
+mod unzip_error;
 
 pub use collection_collision::*;
 pub use collection_error::*;
@@ -11,3 +12,4 @@ pub use collection_error::*;
 pub use exceeds_capacity::*;
 pub use item_count_mismatch::*;
 pub use one_of2::*;
+pub use unzip_error::*;