+++ src/impls/sets/btree_set.rs
@@ -1,6 +1,6 @@
 use std::collections::BTreeSet;
 
-use fluent_result::bool::Expect;
+use fluent_result::bool::dbg::Expect;
 
 use crate::{CollectionCollision, TryExtend, TryExtendSafe, TryFromIterator};