+++ src/utils/mod.rs
@@ -1,13 +1,15 @@
-mod debug_option;
 mod fixed_size_hint;
 mod fold_mut;
 mod identifiable;
+mod no_drop;
+mod option_type_debug;
 
-pub(crate) use debug_option::OptionTypeDebug;
+pub(crate) use no_drop::NoDrop;
+pub(crate) use option_type_debug::OptionTypeDebug;
 
 #[allow(unused_imports)]
-pub use fixed_size_hint::{FixedSizeHint, FixedSizeHintEx};
+pub use fixed_size_hint::*;
 #[allow(unused_imports)]
-pub use fold_mut::FoldMut;
+pub use fold_mut::*;
 #[allow(unused_imports)]
-pub use identifiable::Identifiable;
+pub use identifiable::*;