+++ tests/doc/main.rs
@@ -1,4 +1,5 @@
 mod array;
 mod arrayvec;
 mod result;
+mod try_unzip;
 mod tuples;